requirements.txt
tensorflow==2.16.2
opencv-python==4.11.0.86
matplotlib==3.9.4
numpy==1.26.4
Pillow==11.3.0
requests==2.32.5
pandas==2.3.3
